requirements.txt
lxml
pdfplumber
Pillow
pdf2image
PyPDF2
python-docx
